fix cursor placement bug
This commit fixes the bug where the cursor was placed incorrectly after
completing a tag if the was completed at the very end of the file. It also
fixes a just-discovered bug where multiple tags completed would cause
the tag to be placed incorrectly.

The auto-indent feature was temporarily removed to simplify the debugging
process. Next commits should reintroduce this feature.