reimplement auto-indent feature
This commit reimplements the feature, fixing a bug which resulted in
incorrect placement of the cursor if the tag was not on the 0-indentation
level. It's still probably going to cause problems for someone because
the implementation would make the indenting get recorded in a macro.

The current implementation uses `vis:feedkeys()` to hit `<enter><enter>`
and `<Tab>`. This way respects the user's `expandtab` setting, but makes
the keys get recorded in any macro. An alternative would be to look up
the user's settings and change the indentation that way, but I never
use macros, so I'm not too concerned about it for now.